chore(tests_integration): move node runner test util
commit-id:2959cc90
